individual-detection/requirements.txt
matplotlib~=3.7.1
pandas~=2.0.1
tqdm~=4.65.0
scikit_learn~=1.2.2
opencv-python~=4.7.0.72
hydra-core~=1.3.2
pytest~=7.3.1
mypy~=1.3.0
